% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{compose_get_call}
\alias{compose_get_call}
\title{Title \lifecycle{experimental}}
\usage{
compose_get_call(call_endpoint, call_meta, call_args)
}
\arguments{
\item{call_args}{}
}
\value{

}
\description{
Title \lifecycle{experimental}
}
