% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpeasy-package.R
\docType{package}
\name{httpeasy-package}
\alias{httpeasy}
\alias{httpeasy-package}
\title{httpeasy: Easy peasy HTTP}
\description{
Helps with some low-level HTTP stuff.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/rappster/httpeasy}
}

}
\author{
\strong{Maintainer}: Janko Thyson \email{janko@rappster.io} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
