% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_server.R
\name{start_api_server}
\alias{start_api_server}
\title{Title \lifecycle{experimental}}
\usage{
start_api_server(
  path = confx::conf_get("api_server/script", "config.yml"),
  host = confx::conf_get("api_server/host", "config.yml"),
  port = confx::conf_get("api_server/port", "config.yml")
)
}
\arguments{
\item{port}{}
}
\value{

}
\description{
Title \lifecycle{experimental}
}
