% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_server.R
\name{stop_api_server}
\alias{stop_api_server}
\title{Title \lifecycle{experimental}}
\usage{
stop_api_server(callr_obj)
}
\arguments{
\item{callr_obj}{}
}
\value{

}
\description{
Title \lifecycle{experimental}
}
